refactor: Refactor: split orbit-store v2_bundle.rs into focused submodu… [ORB-00120]
Refactor: split orbit-store v2_bundle.rs into focused submodules

Planned-By: gemini
